Rename option "kdedir" to "install-dir"
The kdedir is ambiguous, it is not clear if it is source, build or install dir.

The kdedir naming dates back to the 2003 era when you had to define KDEDIR to define where KDE 3 was installed, just as QTDIR was used to define the location of the Qt 3 install.

mpyne later added prefix for non-KDE modules and should have switched things over afterwards but never got around to it.

By KDE SC 5 the requirement to use KDEDIR (or as it was by then, KDEDIRS) was dropped in favor of searching PATH.

See https://invent.kde.org/sdk/kdesrc-build/-/commit/2f7b9477b7d44a2f88c148846c8f80f6921fd4ef#note_839381

Original commit: e5c028cd
https://invent.kde.org/sdk/kdesrc-build/-/commit/e5c028cd18fa9f5e508e7b88d64033880eca7fa6