Fix default value for "directory-layout" option
It was silently allowed to have unset "directory-layout" option value, that was treated as "metadata" value.

Previously, it defaulted to "metadata". Then commit 7fe7c940 (BUG: 461279) changed the default value in documentation to "flat", but not in the code.

This commit treats any incorrect value (not one of "flat", "invent", "metadata"), including empty string, as a "flat" value with warning.

Add the forgotten default BuildContext value, and add it in FirstRun.