Don't build ZXing
That's a dependency that practically all distros have in a sufficiently
new version, and which doesn't depend on Qt, so there is no need to build
this yourself. Furthermore, building the master branch of that has caused
problems a few times as that can contain source incompatible changes our
code isn't prepared for yet.

CCBUG: 480662