Support ignoring modules of all scm types
When BuildContext has modules in ignore_list, before actually ignoring it, it was checking if the module's scmType is "proj".

There are other scm types. Currently, they are: "git", "qt5", "proj". The abstract Updater has no type; there is also "metadata" type.

Some modules that are not of type "proj" (Updater::KDEProject), can theoretically be listed as ignored in context, so I remove this check.