KDEProjects: Place metadata always to the same location
Users can have several configurations they run kdesrc-build with (with `--rc-file`, and specifying different `source-dir` and other paths in that file).

This commit makes the metadata module to always be placed in single place. So it would be easier to work with. For example, to debug some situations with metadata changes.