Remove --print-modules and --list-build options
There are already `--query module-set` and `--query branch` options, no need to create unnecessarily entities.

The --list-build tree indentation was not properly implemented, but made correctly in dependency-tree.

The #dependency-level sticky option was never set, so print-modules indentation was broken forever.

The zsh completion script relied on --list-build option and needs to be reworked.