Fix using non-default persistent data file
The `loadPersistentOptions` should be called _after_ the config reading was done. Because config can contain the persistent-data-file option.

Before this commit, it tried to load persistent options from yet unknown location (not yet loaded persistent-data-file option from config), defaulted to standard name, which may not exist.

This also caused that when user renamed his persistent file, and written its name in the persistent-data-file option, then its content got erased, and overwritten by _only_ those modules which were build in the last kdesrc-build invocation.

This is because it tried to read from wrong location, but at write time, it knew the correct location and overwrite the file.