Remove update-shellrc step
It basically just added to the shellrc the path to the running script. That is not handled correctly in all cases.

Also, it tried to add kdesrc-run bash autocompletion, but the code was not handling all cases correctly anyway.

Instead, we will offer to install the file in the correct place in the first place.
Regarding PATH, that should be in user's control.