FirstRun: Avoid the Application taking lock
Creating an instance of Application without parameters creates a lock file. Mark it as todo, and make a hack to avoid taking lock for the current case.

We also cannot use finish method because it also does exit, but in this case we need to continue.

If we return in `if (!$workLoad->{build})`, then when normal using script as `kdesrc-build --metadata-only` (i.e. not from FirstRun), then the script tries to proceed with other operations (in eval block: `my $result = $app->runAllModulePhases();`)

Will rework this later.