FirstRun: Download repo-metadata without reading config
When running --install-distro-packages step, do not read config yet. It has "include" line, that points to metadata module-dependencies. But the metadata may not yet be downloaded. At the same time, normal metadata downloading already expected the normal config, to read the persistent file, to write the last-metadata-update.

We could skip reading config further than global section. But still, user may not even have a config yet. Let the user decide which step he wants first: metadata downloading or generating config.

So we just skip reading config for this case, and not writing last-metadata-update to persistent data file (because user may want to use different name for it rather than default name, and we do not rely on config).