data/perl-dependencies/gentoo.ini: sync with opensuse.ini
This file referred to a package that never existed (I've checked
the historic repo.. not there either) called YAML-XS.  Based on some
blaming, I believe the intention was to depend on JSON-XS.  YAML-PP can
also be swapped for YAML-LibYAML to match other distros.  URI was not
specified, so add it.

Updated package list is based off of the list of packages in
opensuse.ini.