Fix confusion between sysadmin-repo-metadata and sysadmin/repo-metadata
I had a lot of trouble building kwin_x11 because kdesrc-build was
updating ~/.local/state/sysadmin-repo-metadata but then reading
information from ~/.local/state/sysadmin/repo-metadata (in
_resolveModuleDependencyGraph)

The solution is to set the directory-layout for that special module
to the special value "metadata" which was still handled by the code
but was never set.
Very likely due to commit 642bdc2e326f8eae6cbc2a07bfe37fc1044952b9
'Fix default value for "directory-layout" option' which (correctly)
changed the default away from the value "metadata".

This makes everything work with sysadmin/repo-metadata again, as it
used to be. You can delete sysadmin-repo-metadata which was never meant
to exist.